Rename and move _JSONSpeechAPI to HTTPSpeechAPI.
